*: Add some hardening compile options.
Also fix some warnings and potential issues that building the entire EuropaTools codebase under those flags showed.

Later on once we drop the libpixel and stbiw dependencies, we can probably turn on -Werror for release builds.

+++ src/tools/eupak/Utils.cpp
@@ -53,7 +53,7 @@ namespace eupak {
 		auto count = std::strftime(&buf[0], sizeof(buf), format.data(), &tmObject);
 
 		// an error occured, probably.
-		if(count == -1)
+		if(count == static_cast<std::size_t>(-1))
 			return "";
 
 		return { buf, count };
